requirements.txt
PyQRCode
pypng
pyzbar
Pillow